Resolve another passing future
types/records/bharshbarger/remoteBlankRecord used to experience a record being
shallow copied across locales so the "data" field was pointing back at the
original locale instead of the new one, and that access of a local variable from
a remote locale yielded the remote locale id instead of the local id.  Tom says:

"To get remote strings to print correctly, I ensured that references to records
are carried through a chain of calls (except for special cases and cases I have
missed).  In no-local and multilocale runs, only references carry locale
information, so only objects passed by reference will carry along that
information.  Test cases involving records-in-tuples are currently failing in
multilocale for this reason.  Tuples are returned by value, so locality of a
tuple return value is lost.  Some significant work will be required to correct
this."

I choose to leave the skipif in place, since the purpose of the test was to
verify remote behavior.