More passing future fixes
distributions/sungeun/misc/useDummyDist was failing with a function resolution
bug - it couldn't find the type specifier for a distribution when it should be
able to, and now it can.  Tom says:

"In many places, I added calls to getValueType(), to counteract the effect of
passing more objects by reference.  My guess is that somewhere the distribution
was getting passed by reference where the type matching should have been using
the base type."

users/holtbg/tupleOfTupleIteration used to fail assertions when iterating over a
tuple of tuples because the refType field of a type was expected to be filled,
but it didn't need to be because the type itself was already a reference.  Tom
says:

"Since I use ref types more extensively, I probably made some improvements in
passing."