Leak condition variables used for sync vars under cygwin
Some bug results in condvars still being used when we want to destroy them on
cygwin. For now, just leak them to avoid regressions that are a result of the
undefined behavior of trying to destroy an in-use condvar. Post release (or
before in the unlikely event we have some time) we should revisit this and
track down when the condvar is still being used.

This should resolve most of the sporadic timeouts and failures under cygwin,
though test_cholesky_performance will still fail.