Improve application of FLAG_AUTO_DESTROY_FN in compiler
This commit just adds the flag FLAG_AUTO_DESTROY_FN to all autoDestroy functions
that are mapped to types in function resolution.  It gets around what I
observed in the module code, which was that some chpl__autoDestroy functions
had the pragma applied to them and some did not.  I believe that if this change
works, we will not need the pragma in the module code any more either, which
would be nice, and that a similar change could be done for FLAG_AUTO_COPY_FN.

I encountered this when investigating a change on the string-as-rec branch.
The problem is likely also visible on master, and this change could lead to
removing the pragma version of the flag and its application in module code, so
I figured it was a good thing independent of its use to me in string-as-rec.