Add optimization to remove extra autoCopy/autoDestroy pairs
Will only fire when a variable is defined once and only used in one autoCopy
and one autoDestroy.  If a variable has an additional use, it will not trigger
this optimization.  Also, if the variable's definition is as an out argument to
a function call, this optimization will not be triggered - I decided to start
as conservatively as possible.  Once a variable has been determined as needing
this optimization, its autoDestroy function will be removed and its use in the
autoCopy function will be replaced by the right hand side of its def expression.
If the right hand side is itself an autoCopy, then instead of pasting the pure
autoCopy call into the use it will be peeled off and the innards of the def's
autoCopy will be inserted into the use's autoCopy.