Remove pragma "auto destroy fn" from module and test code, and remove pragmaness
Since I added this flag during resolution, these functions no longer needed to
be labeled with a pragma, they will be recognized as autoDestroys on their own.
With those removed, I disabled the ability for the flag to be added via a pragma
(since it wasn't applied to anything beyond chpl__autoDestroy()s in the module
code).