Move most task callback support into common code.
The bulk of task callback support, minus common type definitions and
such, had been implemented separately in tasks=fifo and =qthreads.  Here
we move most of it into common code.  The functions that users call to
install and uninstall callbacks, and the one the tasking layers call to
actually do the callbacks, are in chpl-tasks-callbacks.c.  Definitions
for the latter and wrappers that can be inlined to reduce overhead are
in chpl-tasks-callbacks-internal.h.  All of the callback-related code
except the calls to actually do the callbacks is removed from the
qthreads and fifo tasking layers.