Add a comment clarifying the case that was handled and clean up a little
Mike suggested that the comments would be clearer if an example of the code
being cleaned up was provided.  He also suggested adding a note that the
code is intended to solve a problem that we hope will not exist in the future.

While here, I condensed the code a little - the else branch for other types of
defs was not necessary, so I removed it and moved the code that followed it
into the autoCopy def case.