Initialize "large" numeric arrays in parallel by default
Do parallel initialization for "large" numeric arrays. Large being at least
2MB, which was chosen experimentally by comparing the performance of serial vs.
parallel initialization on a 2 core laptop, 8 core server, and 24 core XC40.
Long term we want users to be able to specify how an array should be
initialized. In the short term, we want users to be able to get better first
touch behavior out of arrays. As a parallel language we believe it's safe to
assume that most large arrays will be used in parallel, and thus defaulting to
parallel initialization will usually result in better first touch behavior. For
the stream benchmark, this results in a 2x performance improvement.

The reason this is limited to numeric types is because we ran into issues
trying to parallel initialize arrays of arrays and thus arrays of aggregate
types where at least one field is an array. There's more details in init_elts,
but the problem is basically that domains store a linked list of their arrays
and removal became expensive for arrays of arrays. Removal from a linked list
is O(n). However, when we were doing serial initialization we happened to be
adding and removing arrays from the domain's list in the same order. Removal
was effectively O(1), since the array to remove was always at the beginning of
the list. Ideally the domain's arrays would be stored as an associative domain
or something with < log(n) add/find/remove times. We can't use our associative
domain as-is because the list of arrays is part of the base domain so we'd have
a circular reference. We should be able to peel most of the associative domain
code into a standalone data structure that the associative domain would
forward to. I think we also have to update the default hash for arrays as the
current computation is expensive and I believe default initialized arrays would
hash to the same value.

A stepping stone towards initializing all types in parallel would be to do
parallel initialization for plain old data (POD) types. The threshold for
whether parallel initialization is worthwhile might have to be updated since we
also have to account for the construction cost of the records. Currently the
isPOD functionality lives on the string-as-rec branch and it wasn't trivial to
bring into master.