Fixed placement of some autoDestroys
The insertAutoCopyAutoDestroy pass relies on a simple restructuring
of each function's body, to facilitates ownership analysis
in the presence of early returns.

Namely, the bulk of the function body is moved into a newly-created
nested scope. I.e. the function body like this:

  {
    DefExpr ret;
    STATEMENTS;
    DefExpr _end_PROCNAME;
    return ret;
  }

is restructured to:

  {
    DefExpr ret;
    {
      STATEMENTS;
    }
    DefExpr _end_PROCNAME;
    return ret;
  }

(Implementation-wise, the original function-body BlockStmt is retained,
and a new BlockStmt for the outer scope is created.)

The analysis used to produce incorrect results e.g. in these tests:

 modules/standard/FileSystem/lydia/copy/copying
 modules/standard/FileSystem/lydia/copyFile/copyingContents
 modules/standard/FileSystem/lydia/copyTree/copying
 modules/standard/FileSystem/lydia/copyTree/no_src
 modules/standard/FileSystem/lydia/isMount/notMount

for this function:

  proc isDir(out error:syserr, name:string):bool {
    ...;
    if (...) {
      return false;
    }
    ...
    return ...;
  }

There, there was a string variable (I think it was a temp)
that was declared+initialized after the if-statement.
The autoDestroy() for that variable was placed after the end label
instead of before that label. So when the if-condition was true,
that autoDestroy() was executed without that variable getting
initialized first. That caused accesses to uninitialized memory.

This commit makes a couple of changes to get the analysis
to produce correct results in this case and another related case.
Both changes were suggested by Tom H.

* Search for the return label - do not give up when it is not
the second-last statement in the function body.

* Also move the formal temps' DefExprs into the outer scope
next the return variable's DefExpr.

While there, I eliminated a collectDefExprs() and a collect_stmts().
Instead, traverse the statements in the BlockStmt's AList directly.