Do not let local var outlive its scope in 'begin'
The uts_deq.chpl benchmark had this code since its inception
(using then-appropriate syntax):

  proc balance_load ... {
    ...
	var work = ...; // of a record type
    ...
	begin with (ref work) create_tree(work);
    ...
  }

Now that we do not want local variables to outlive their lexical scope
in begin-statements, I am replacing the 'ref' task intent with 'in'.
This works for the intention of the code, which is to hand off
'work' to create_tree(), and ignore the result.

While there, replaced 'inout' intents with 'ref'.
'inout' intents had been there since the code was first written.
I imagine 'inout' was used because 'ref' was not available on records
back then.