Replace sync with atomic variables.
While I was looking at this code, I upgraded it from syncs to atomics.
This made it more elegant/short, hopefully faster, and
replaced a 'single' variable with a call to waitFor().

I needed an implementation of atomic max, which I copied from Ben's fannkuch:
  proc atomic_int64.max(other: int)
It probably belongs in modules/internal/Atomics.chpl.