Added a new dsiEqualDMaps() interface to Block and DefaultDist
It seems we've never supported the ability to check equality
of two domain maps, which is something we now need as we add
support for formal domain type checking.  Here, I've added
implementations of the supporting dsi routine for Block and
DefaultDist.  Other distributions will need this as well (a
TODO).