Add a new test implementing proposed formal-actual domain checks
This is a test that I've been using to develop the new interpretation
of the domains of formal array arguments.  One thing that's bothersome
about it is that the errors are reported in terms of the callee
function rather than the caller, which is not particularly helpful.
I haven't figured out how to deal with this yet, though...