Place polling threads on the last available CPU.
Putting the polling thread in a predictable place may help the tasking
layer avoid interference from it.  In particular, the current tasking
layers that do place their threads put them by default on cores starting
at the first available one, so this new polling thread placement should
maximize how many hardware threads we can use for hosting tasks before
we start to see interference from the polling thread.  While this is a
bit of a simplification, nevertheless the new placement seems like it
should always be as good or better than leaving the polling thread
unplaced.

The new chplsys:chpl_moveToLastCPU() function is a bit of a hack.  In an
ideal world we would use something like hwloc for this.  But hwloc is
not yet a first-class package for us, only an adjunct to qthreads, so we
can't depend on its presence.  And the affinity calls we're using are
only available for (non-MIC) Linux.  It's best to hide all this detail
in chplsys, and only have one copy of it, instead of putting it out in
public in the tasking layers.