Further changes to this test to get it working in new world
This test originally was written as a number of cases following
the following pattern:

    const TD1D: domain(1) = Space1.translate(-o5);
    var TA1D: [TD1D] int;
    ...foo(Dom1D, TA1D, TA1D)...

which relied on the formal argument of foo() to cause reindexing
to take place:

    proc foo(TD: domain, A: [TD] int, TA) {

Under the new "formal array arguments with domains" interpretation
in which reindexing is no longer done, I had originally rewritten
these calls as:

    foo(Dom1D, TA1D.reindex(Dom1D), TA1D)

Basically "do the reindexing manually, using the domain that we're
going to pass in as argument 1."  However, it turns out that now
that I have the type checking implemented, it no longer works as
I thought it would for the following reason:

* Space1 is a DefaultRectangular
* TD1D therefore also is
* TA1D therefore also is
* Dom1D is, however, a distributed domain
* Yet, TA1D.reindex(Dom1D) is DefaultRectangular as well because
  reindexing maintains the distribution/domain map of the original
  array, not the one that's setting the new index set (to permit
  default rectangular domains to slice/reindex/etc. distributed stuff
  without localizing it all).

So therefore, foo() complains (as it should) because it's being passed
a Block distributed domain and an array that is not block distributed
as the declaration suggests it should be.

Reading the first commit log for this test, it seems that this
non-distributed reindexing was intentional (i.e., "What happens when a
non-distributed thing is indexed by a distributed thing?  Does it
work?"), and variations on this test seem to handle the case that I
was expecting (both array and reindexing domain being distributed).

So, the fix I took was to create an explicitly non-distributed version
of the first argument to end up with a DefaultRectangular that would
match the reindexed TA1D.  This starts to make the test a little
funky (use these distributed things to get nondistributed things), but
still seems worthwhile to keep.  Thus, the new calls look like:

    foo({(...Dom1D.dims())}, TA1D.reindex(Dom1D), TA1D)

where {(...D.dims())} is my favorite sugar for converting an existing
domain to a default rectangular regardless of what its original domain
map happened to be.