Followed up on review comments
* removed "index<0" tests as index is size_t which is unsigned

* annotated debugging functions with 'const'

As to 'const': I was able to add it to only a couple of functions.
The remaining functions either invoke BasicBlock::printBitVectorSets
or do this.symbolIndex[sym].

BasicBlock::printBitVectorSets cannot be easily marked 'const'
because it invokes for_vector.

I don't know why I can't use [] on a const vector symbolIndex.