Fixed a compiler failure in autoCopy.chpl
The code in OwnershipFlowManager::insertAutoDestroys() does this:

...........
  // Move away the return statement so insertAtTail() do not add after it.
  CallExpr* ret = removeReturnStmt(_fn);

  // and then destroy each symbol at the end of its containing scope.
  insertAutoDestroy(&to_cons);

  // Reinsert the return statement.
  if (ret) _fn->body->insertAtTail(ret);
...........

This temporary removal of the return statement is done so that this code
in OwnershipFlowManager::insertAutoDestroyAtScopeExit(Symbol* sym):

...........
  BlockStmt* scope = sym->getDeclarationScope();
  scope->insertAtTail(autoDestroyCall);
...........

does not insert *after* the return statement.

That same function also contained _fn->getReturnSymbol()
which does *not* work without the return statement.

As a workaround, I stash away the return symbol in the OwnershipFlowManager
itself, in its constructor.

Note that the only test where this fires is:

  test/performance/sungeun/multilocale/autoCopy.chpl

because of:

...........
pragma "return value is not owned"
proc autoCopy(x) { return chpl__autoCopy(x); }
...........

See commit 325b0894 for details on this.