Add an array equality method
The == operator on arrays returns an array of bools, and while we've
gone back and forth over time about whether or not this is the right
interpretation (see mail archives), we haven't ever changed it.

Here, I'm adding an equals method to arrays that checks whether
they're equal in the sense of (a) having the same rank, (b)
if so, having the same size/shape, (c) if so, having all their
elements == using a reduction.