Replace a hand reduction for equality with array.equals()
Realized that the simple '& reduce (A == B)' pattern that I used for
array equality is problematic in that (a) it doesn't check to see if
the two things can be zipper-promoted like this and (b) because of
the mismatched size bug we have in zippered promotion, if A is bigger
than B, we would get a false positive.

For this reason, introduced the _array.equals() method in my
previous commit to ChapelArray.chpl and use it here instead.

Also changed the default case to a param since it always returns
false.