Added equality support functions for BlockCyclic, Cyclic, Replicated
This adds the dsi routines necessary for these three distributions to
do equality checks on their distribution values.  The one for
BlockCyclic is arguably shallow because rather than answering "are
these two distributions the same?" it's answering "are they
parameterized by the same values?" which is not as interesting.  See
the TODO comment for more details.  The fix is probably easy, modulo
the C negative modulo issue which Chapel inherits...

While here, I also stumbled across an issue in which the
ReplicatedDist blows up if handed a multidimensional target locales
array, so I put in a compiler error for this case.