Prevent addition of FLAG_NOT_POD to classes during scopeResolve
Classes are considered POD types, but if a class defined an initialize function,
it was getting marked as not POD during scopeResolve.  This caused "ignore
noinit" to be applied to it, and that caused the defaultValue for classes to
be given to return variables of that type during function resolution.  However,
the defaultValue field for classes is of nilType, and we can't assign to a
nilType from a different type (the defaultValue field is deprecated in favor pf
_defaultOf functions), causing the compiler error we were seeing for the
madness tests.  With this change, the class will again be considered a POD type.

I'll commit the code to remove the defaultValue assignment after I'm certain
this part works, to see if that can also be removed.  It wasn't present on
master.