Revert "Removed unused conditions in flattenFunctions.cpp"
This reverts commit 67bc3f8dc497b8318504b6ef2235a7f6c57f291c

Turns out that the removed checks DID fire in this test:

  domains/sungeun/sparse/index_not_in_domain_1

on this expr:

  (894668 CallExpr move
    (894669 SymExpr 'p_i_3[804499]:int(64)[10]')
    (894670 SymExpr 'chpl__saIdxCopy[894663]:int(64)[10]'))

and their removal caused a valgrind failure in that test

I am putting these checks back in for now.

On string-as-rec, we still need to either fix those checks to handle
reference temps - or find a way not to use them.
What were these checks doing in the first place?