Make comm pthread arg static; won't be referenced until long from now.
Silly me.  The argument block for the comm pthread has to outlast the
activation of chpl_task_createCommTask(), since it won't be referenced
until the comm pthread actually calls the wrapper.  That being so, make
it static.  (This is safe because it's only used precisely once on each
node).  While here, give it its own type.  Reusing the one for qthread
wrappers is inelegant.