Removed unused conditions in flattenFunctions.cpp
I am re-introducing, on string-as-rec, the PR #2362 reverted in #2436
// corresponding commits:  67bc3f8 and 5758082

The motivation for removing those checks (see #2362) was that
on string-as-rec they would be wrong because reference temps are now
inserted earlier and these checks would not fire in their presence.

However, turns out that these checks are needed in this test:

  domains/sungeun/sparse/index_not_in_domain_1.chpl

where valgrind reports an error when the checks are removed.
So I re-introduced these checks back on master (#2436)
so valgrind succeeds.

However, for string-as-rec we need to investigate whether these checks
are working correctly. Towards that end, I am again removing them.
The above test continue failing under valgrind for now,
and I will add a JIRA issue for us to come back and investigate.