Refactor the Barrier standard module into a single main class
The Barrier class is now the single type for barriers in this module, but it is
implemented using either atomics or syncs chosen with an enum at construction
time.

Updated the tests that use this module to choose the same version they had been
using (atomics or syncs) and to delete the class instances when they are done
with them and to remove "with (ref bar)" clauses from coforall loops.

test/parallel/taskPar/sungeun/barrier/commDiags.chpl is currently regressing
for comm-gasnet due to a few extra gets.