Add a new flag --[no-]formal-domain-checks to disable new checks
Added a new flag --no-formal-domain-checks that disables the
new checking of domains between formal array arguments that
specify them and actual arguments.  This commit adds the flag
but does not yet make use of it.

Updated the man page to describe the flag.

Since --fast implies --no-checks, made the routine which sets
all the --fast settings call the routine that sets all the
--no-checks settings for simplicity (previously, the changes
were duplicated across the two functions which seemed surprising.