Add param argument indicating whether or not runtime checks are desired
This commit implements the --formal-domain-checks flag by passing a
param bool into the verification function indicating whether or not
the runtime checks should be done.  Note that the static checks are
done in either event.

I also made the check routine inline so that, when --no-formal-domain-checks
was thrown, there should be zero overhead in terms of generated code
for such routines.