Use "is pod type" primitive when checking for bulk transfer support
Adds new functions to Types.chpl that query the "is pod type" primitive.
These are "no doc"ed for now, but could be exposed if we want to. We
should probably define what a POD type is in the spec before making
these functions public.

isPODType is now used to check if a given record or tuple type can
support bulk transfer.