Allow promotion wrappers to support fast followers
For promoted operations we build up wrapper serial, follower, and leader
iterator functions. These wrappers just invoke the iterators for the type being
promoted and apply whatever operation is being promoted. Since these are really
just wrapper or forwarding iterators, we need to be able to convey whether fast
followers were requested by the callee. To do that we just add a "fast" param
to the follower wrapper and if it's set call "_toFastFollower*" instead of
"_toFollower*"

Note that fast followers will not actually be invoked yet, because there are no
static or dynamic fast follower checks that the callee uses to determine if
it's safe, so it falls back to the default false cases.

This also removes 2 "_toFastFollower" functions in the modules that were
defined for _iteratorClass and _iteratorRecord.

I also added a few minor comments to the wrapper building code, since it took
me so long to wrap my head around it.