Enabled reduce intents with foralls over domains
This is done by handling the case where the parallel iterator
(leader or standalone) has yields outside of a parallel construct.

The shortcoming of this PR is that it does not handle properly the case
observed in DefaultRectangular array standalone and leader iterators.
They have a yield is inside a 'for' loop over a parallel iterator,
not a serial iterator. Since 'xShadow' below is shared across iterations
of the for loop, it will be accessed in parallel, leading to data races.

IMPLEMENTATION DETAILS

Reduce intents were introduced in #1386, primarily in commit 16880ce.
There, the code:

```
    var x = 5;
    forall idx in myIterator(args) with (+ reduce x) do
      BODY(x);
    writeln(x);
```

assuming it invokes the standlone myIterator(),
gets rewritten to:

```
    var xOuter = 5;
    const globalOp = new SumReduceScanOp(xOuter.type);
    for (idx, ref xShadow) in myIterator(args, globalOp, iterKind.standalone) {
      BODY(xShadow);  // substitute xShadow for x
    }
    x = globalOp.generate();
    writeln(x);
```

At the same time, the standalone iterator:

```
    iter myIterator(args, param tag) where tag == iterKind.standalone {
      ... yield 6; ...
    }
```

is be rewritten to:

```
    iter myIterator(args, parentOp, param tag) ... {
      ...
      var xShadow ...;
      yield (6, ref xShadow);
      ...
    }
```

#1386 handled only the case where yields were inside a task construct.

This PR extends the above scheme to handle yields in the parallel iterator
that are not in any task constructs. It works whether or not the iterator
also has task constructs, with or without yields in them.

This PR's change is this. If the above "yield 6;" is not in any task
construct, before the compiler would issue USR_FATAL.
Now, the above iterator is instead rewritten like so:

```
    iter myIterator(args, parentOp, param tag) ... {
      var xShadow = parentOp.identity;
      ...
      yield (6, ref xShadow);
      ...
      parentOp.accumulate(xShadow);
    }
```

This handles forall loops over domains, as the doman parallel iterators
have yields outside task constructs.

ADDITIONAL TEST

I also added a test, ri-4-arrdom.chpl, which tests such foralls
with reduce intents. It is also a stress test like ri-3-stress.chpl.
It also contains foralls over arrays, although those are commented out
because I do not expect them to work.

  test/parallel/forall/vass/ri-4-arrdom.chpl