Rework noinit warning to point to use, not type's def
These futures failed because the line numbers in ChapelArray.chpl
changed, causing the declaration point of arrays and domains to
change.  Looking at the code, it seems to me that one would prefer
the warnings to point to the point where 'noinit' was used (since
that's where the user's code was wrong and since the type name is
printed out) rather than where the type was defined (which has no
bearing on where its noinit function may or may not have been
defined).  This also has the side benefit of making these futures
more maintainable as ChapelArray.chpl continues to evolve, as it
will.