Add some comments to Barrier standard module
Added a comment to the future direction section explaining that the Barrier
type will probably become a record instead of a class in the future.  Also
explain that when this happens the only code change that should be required
for Barrier clients is that they can remove explicit deletes.

Fixed a typo in another comment.