Change parser + AST/build.cpp to accept 'require'
This change adds support for 'require "foo.h"' to the grammar and AST
builder.  I preserved the 'use "foo.h"' form with a warning to ease
transition for any users who are making use of the old form, though
it should be retired after 1.12 goes out.

This commit includes one new test to lock in the warning for the
deprecated behavior.