Update error location for "working" noinit test case
Follow-up to PR #2455 (i.e., "This is why you shouldn't fix
regressions before you've fully woken up in the morning."  :)

PR #2455 improved the location of noinit error messages this morning
to improve some failing futures, but in writing it, I failed to do a
full test or grep to see whether other, possibly passing, tests would
be affected/improved by the change.  Turns out that there are --
whoops.

This commit updates the location of the warning in this test to
similarly point to the noinit.  As in the previous cases, I think this
is an improvement for the user.