Trigger fast followers for promoted expressions
In order for fast followers to trigger the `chpl__staticFastFollowCheck` and
`chpl__dynamicFastFollowCheck` functions must be implemented for the follower
types. For promoted expressions the type is an iterator record which did not
have the above functions.

We needed some way to get the fields of an iterator record in order to check if
they were candidates for fast followers. This commit basically creates a tuple
that contains each value of an iterator record and calls the corresponding fast
follower check on that tuple. In other words we try to convert the iterator
record to a tuple since we already have fast follower check functions for
tuples.

However, when promotion wrappers are created during function resolution the
iterator record is empty and will only be filled in during iterator inlining. To
get around this a new primitive (PRIM_ITERATOR_RECORD_FIELD_VALUE_BY_FORMAL) was
introduced. The primitive takes an iterator record and the formal that will
later be transformed into a field. The formal is used to determine the type.
During iterator inlining the iterator records fields are finally added, one
for each formal of the wrapper iterator. At this time, we now replace out
primitive with a PRIM_GET_MEMBER_VALUE and replace the formal with the iterator
record's field.

The end result is that we're able to use fast followers for promoted
expressions. This gets us the same performance for promoted version of stream
as we get for the global version.

I still have some cleanup to do and some tests to add, but I'm ran out of time
for feature freeze, so that'll have to wait until code freeze.