Add performance testing on the chapcs machines
Updated nightly and related scripts to check for a new env var,
CHPL_TEST_PERF_CONFIG_NAME. If this is set it will be used instead of
the hostname to indicate what is being run for performance tests. This
will change which paths dat files are placed into for example.