Change release nbody to use 'ref' variables
Previously we were using a nested procedure with 'ref' intents to work around
the lack of 'ref' variables at the time this version was written.

It'll be interesting to see whether this hurts the performance of this version.
Currently, there's a gap on the shootout machine between the "ref var version"
(nbody-blc-ref.chpl) and the release version, though there are other differences
than this one.