Fix problems with Spawn with GASNet segment=fast
With GASNet and segment=fast, the Chapel heap is not
necessarily available to a forked process. As a result,
if the command, arguments, and environment for the spawned
process are allocated on the Chapel heap, the spawn
will fail.

In this patch, I added some simple support functions in
qio_popen.h/qio_popen.c that use the C allocator. I modified
the Spawn.chpl module to use these support functions.

While there, I switched from using c_string_copy to c_string
since c_string_copy is not strictly necessary and it
is removed on the string-as-rec branch.