Remove extraneous warning for scope resolution of private symbols
When I had initially implemented private symbols, I thought it would be useful
to generate a warning message if we were ignoring a private symbol during
scope resolution.  The case this would catch would be when a user is 'use'ing
a module A and trying to access its private symbol 'foo', and due to another
module at a further distance (say, B, which is used by A or used by another
module their program uses) this compiles but because B.foo is different than
A.foo, the user experiences unexpected behavior without a warning on our part.

In the case where the public 'foo' was intended, this warning message would be
quite annoying.  I had debated about making this warning opt-in, but have been
convinced by others to just remove the message entirely.