Update chplenv compiler version comparisons
A while back we changed get_compiler_version to return a named tuple instead of
a float because floating point comparison didn't work correctly for things like
comparing 4.1 to 4.10. This updates get_compiler_version to include the
revision and build in the tuple returned just in case we need it in the future.
It also moves the regex logic that was in get_compiler_version() into
CompVersion() so clients can use it to compare against values returned by
get_compiler_version()

This converts comparisons against major and then minor number to just compare
against a compiler version tuple.

Things like:

  'if version.major > 4 or (version.major == 4 and version.minor >= 1):'

are now just:

  "if version >= CompVersion('4.1'):"

which I think is much easier to read. You could also just do:

  "if version >= (4, 1, 0):"

but I don't think this is as easy to read

I need to make some changes to the chplenv scripts that will need the cce
version number. I've had this change in a branch for a while and figured this
is a good time to pull it into master since it'll make the cce version number
comparison easier.