Add support for 'cray' to the get_compiler_version utility
Add ability to ask for the cray compiler version in the chplenv scripts. We use
`cc -V` to get version information and then parse that for the version string.
We believe the format of `cc -V` has been consistent for decades and don't see
it changing anytime soon so this should be safe. An upcoming commit to enable
qthreads for CEE 8.4 and newer will require this functionality.

The output of `cc -V` is:
  "<product name> Version <version number> <current date and time>"