Default to qthreads over muxed
ugni+qthreads performance has steadily been improving and for most cases is
better than or the same as ugni+muxed. The perf team and Greg have agreed that
it's time to use qthreads over muxed, especially since we'd like to retire
muxed.

Muxed is still being built into the module and tested, it just won't be the
default.