Get cray version number from CRAY_CC_VERSION instead of `cc -V`
Using `cc -V` is the ideal way we'd like to get the version number but
unfortunately it adds an extra subprocess call to our chplenv scripts which
increases the compilation time of every chapel program. Additionally `cc -V`
itself is pretty slow.

We believe CRAY_CC_VERSION is a pretty reliable way to get the version number
even though it's not documented. We figure that we'll be some of the first
people to know if it's no longer available in future releases so it's not a big
deal if it's removed.