Use qthreads w/o hwloc with cce when we can
Default to qthreads for cce if we have 8.4 or newer or if the module is being
used. CCE 8.4 is required to build qthreads since it has inline assembly.

We build the module with 8.4 so we know we can use qthreads if the module is
being used even if an older cce is loaded.

hwloc doesn't build with CCE 8.4, so turn it off for now. Once CCE 8.4 is
released I'll file some bug reports with the hwloc team.