Add note to STATUS.devel on places where private/public can't be used yet
Since the spec has defined what private and public will mean for type aliases,
but we haven't implemented it yet, add a note to the STATUS.devel file
indicating this.  Also note that it can't be applied to enum declarations,
though I haven't updated the spec for that yet.  The spec has a note in it
already for classes/records and private/public.