Fixed a compiler crash wtih reduce intents
The compiler used to crash on two simple programs courtesy Brad,
which are being added as test/parallel/forall/vass/ri-5-[ab].{chpl,good}

In some cases we used to clone a wrapper and an iterator,
modify them, then re-resolve the iterator and not the wrapper.
I added code to un-do some resolved things in the wrapper
so it can be properly re-resolved.