Change hplx to do reindexing again
Based on the .future, I thought it would be OK to take reindexing
out of this test without affecting its behavior, but we've seen a
lot of passes of it since I did so.  So I'm putting the implicit
reindexing that was done previously back in explicitly, hoping to
make it fail more consistently again.